Apply pattern of an index and a view component further.
